vmmon: avoid set_fs() in HostIFFastClockThread()
The set_fs() function is planned to be dropped from internal kernel API in
5.10 development cycle. It's not clear what purpose does set_fs(KERNEL_DS)
serve in HostIFFastClockThread(); it was certainly needed for the old code
using functions expecting user pointers to access /dev/rtc but that code
has been dropped between Workstation 12.5.7 and 14.0.0 in September 2017.

Drop the temporary set_fs(KERNEL_DS) trick as there doesn't seem to be
a need for it any more.