vmmon: avoid set_fs() in HostIF_SemaphoreWait() and HostIF_SemaphoreSignal()
The set_fs() function is planned to be dropped from internal kernel API in
5.10 development cycle. In HostIF_SemaphoreWait() it is used to allow
calling read() file op on a kernel pointer, in HostIF_SemaphoreSignal() to
allow calling write() file op on a kernel pointer. This is only the case
for pre-5.8 kernels as we need to use kernel_read() and kernel_write()
helpers on 5.8 and newer anyway.

Use kernel_read() and kernel_write() always, like Workstation 16.0.0 does,
so that we no longer need the set_fs(KERNEL_DS) trick. Make the code of
both functions look exactly like in Workstation 16.0.0 to make future
backports easier.